"Red" -> "Ray-mmd" in title & window size adjustments
default window title was still "Red's material tool". Changed that to be consistent with changes made elsewhere. 

Removed minimum and maximum size and added ResizeMode=CanMinimize tag. May want to be reversed later if some kind of scaling is implemented but for now it's unnecessary.

Also changed width from 589 to 590. 589 seems like an awkward number.